third_party/WebKit/LayoutTests/http/tests/inspector/console-show-all-messages.html

<html>
<head>
<script src="console-test.js"></script>
<script src="inspector-test.js"></script>
<script>
console.log("message from page!");

function test()
{
    var checkbox = Console.ConsoleView.instance()._filter._showTargetMessagesCheckbox.inputElement;

    //we can't use usual InspectorTest.dumpConsoleMessages(), because it dumps url of message and it flakes in case of iframe
    function dumpVisibleConsoleMessageText()
    {
        var messageViews = Console.ConsoleView.instance()._visibleViewMessages;
        for (var i = 0; i < messageViews.length; ++i) {
            InspectorTest.addResult(messageViews[i].consoleMessage().messageText);
        }
    }

    InspectorTest.runTestSuite([

            function testInitialState(next)
            {
                if (checkbox.checked)
                    InspectorTest.addResult("\"Show target messages\" checkbox should be unchecked by default");
                dumpVisibleConsoleMessageText();
                next();
            },

            function testPageOnlyMessages(next)
            {
                InspectorTest.changeExecutionContext("top");
                checkbox.click();
                dumpVisibleConsoleMessageText();
                next();
            },

            function testFrameOnlyMessages(next)
            {
                InspectorTest.changeExecutionContext("myIFrame");
                dumpVisibleConsoleMessageText();
                next();
            },

            function testAllMessagesWithFrameContext(next)
            {
                checkbox.click();
                dumpVisibleConsoleMessageText();
                next();
            }
    ]);
}

</script>
</head>

<body>
<p>
Tests that console shows messages only from specific context when show target checkbox is checked.</a>
</p>
<iframe name="myIFrame" src="resources/console-show-all-messages-iframe.html" onload="runTest()"></iframe>

</body>
</html>
